+++ i2cpy/__init__.py
@@ -5,12 +5,14 @@ respectively.
 I2C objects are created attached to a specific bus. They can be initialized
 when created, or initialized later on.
 
-This library is designed to support different I2C driver implementations. At
-present below drivers are supported:
+This library is designed to support different I2C driver implementations.
+At present below drivers are supported:
 
-* CH341 (CH341A, etc)
+* `CH341 <https://www.wch-ic.com/downloads/CH341DS1_PDF.html>`_
+* `CH347 <https://www.wch-ic.com/downloads/CH344DS1_PDF.html>`_
 
-The interface is similar to that of MicroPython's `machine.I2C <https://docs.micropython.org/en/latest/library/machine.I2C.html>`_
+The interface is similar to that of MicroPython's `machine.I2C
+<https://docs.micropython.org/en/latest/library/machine.I2C.html>`_
 
 Example usage:
 
@@ -85,7 +87,7 @@ class I2C:
             module name shipped with this library. For example "foo" means module
             "i2cpy.driver.foo".
             If not specified, it looks at environment variable "I2CPY_DRIVER".
-            And if that's not defined or empty, it finally falls back to "ch341".
+            And if that's not defined or is empty, it finally falls back to "ch341".
         :param auto_init: Call `init()` on object initialization, defaults to True
         """
         self.index = id
@@ -200,6 +202,9 @@ class I2C:
         and return a list of those that respond.
         A device responds if it pulls the SDA line low after its address
         (including a write bit) is sent on the bus.
+
+        Depending on the specific driver and OS platform, scan() may or may not
+        work.
 
         :param start: start address, defaults to 0x08
         :param stop: stop address, defaults to 0x77