+++ i2cpy/driver/ch341/__init__.py
@@ -74,11 +74,12 @@ class CH341(I2CDriverBase):
         else:
             self._init_posix()
 
+        # set baudrate
+        ret = ch341dll.CH341SetStream(self._fd, self.baudrate.value)
+        self._check_ret(ret, "CH341SetStream")
+
     def _init_nt(self):
-        if ch341dll.CH341OpenDevice(self._fd) != -1:
-            ret = ch341dll.CH341SetStream(self._fd, self.baudrate.value)
-            self._check_ret(ret, "CH341SetStream")
-        else:
+        if ch341dll.CH341OpenDevice(self._fd) < 0:
             raise I2COperationFailedError("CH341OpenDevice")
 
     def _init_posix(self):
@@ -93,10 +94,7 @@ class CH341(I2CDriverBase):
             get_chip_version = getattr(ch341dll, "CH34x_GetChipVersion")
             if get_chip_version:
                 chip_ver = (c_uint8 * 1)()
-                ret = ch341dll.CH34x_GetChipVersion(self._fd, chip_ver)
-
-            ret = ch341dll.CH34xSetStream(self._fd, self.baudrate.value)
-            self._check_ret(ret, "CH34xSetStream")
+                ch341dll.CH34x_GetChipVersion(self._fd, chip_ver)
         else:
             raise I2COperationFailedError(
                 "CH341OpenDevice(%s) failed!" % self.device_path